# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CineAI API
  description: Comprehensive movie recommendation API with AI-powered features
  version: 1.0.0
  contact:
    name: CineAI Team
    email: contact@cineai.app
  license:
    name: MIT
servers:
  - url: https://cineai.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Movie:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        overview:
          type: string
        posterPath:
          type: string
        backdropPath:
          type: string
        releaseDate:
          type: string
          format: date
        runtime:
          type: integer
        rating:
          type: number
          format: float
        voteCount:
          type: integer
        genres:
          type: array
          items:
            type: string
        cast:
          type: array
          items:
            type: object
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
...
paths:
...
  /auth/verify-otp:
    post:
      summary: Verify OTP and authenticate
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                token:
                  type: string
                  pattern: '^[0-9]{6}$'
              required:
                - email
                - token
      responses:
        '200':
          description: Authentication successful
        '400':
          $ref: '#/components/responses/ValidationError'
  /movies:
    get:
      summary: Get movies with optional AI recommendations
      tags:
        - Movies
      parameters:
        - name: smart
          in: query
          schema:
            type: boolean
          description: Enable AI-powered recommendations
        - name: behavioral
          in: query
          schema:
            type: boolean
          description: Use behavioral analysis
        - name: realtime
          in: query
          schema:
            type: boolean
          description: Fetch live TMDB data
        - name: query
          in: query
          schema:
            type: string
          description: Search query
        - name: mood
          in: query
          schema:
            type: string
          description: Mood-based filtering
        - name: genres
          in: query
          schema:
            type: string
          description: Comma-separated genre IDs
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: Movies retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      movies:
                        type: array
                        items:
                          $ref: '#/components/schemas/Movie'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
                          hasNext:
                            type: boolean
...
  /ai/chat:
    post:
      summary: AI chat interface
      tags:
        - AI & Recommendations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  example: "Recommend me some horror movies"
                sessionId:
                  type: string
                  format: uuid
                stream:
                  type: boolean
                  default: false
              required:
                - message
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      response:
                        type: string
                      extractedPreferences:
                        type: object
                      sessionId:
                        type: string
...
security:
  - bearerAuth: []